refactor: move locate cleanup into the brokerage transaction handler
- new SanitizeOrderProperties pass runs on every submit, replacing the pre order checks call
- the pass edits a copy of the order properties and hands it to the order; the caller's own instance stays untouched
- send a one-time warning when a locate is removed from a non short order
- move the locate cleanup tests to the transaction handler tests
