+++ Tests/Algorithm/AlgorithmTradingTests.cs
@@ -1892,127 +1892,6 @@ namespace QuantConnect.Tests.Algorithm
             return hashSufficientBuyingPower.IsSufficient;
         }
 
-        // A locate belongs only on an order that opens a short position. Each case: the order
-        // properties carrying the locate, the holdings before the order, the order quantity, and
-        // whether the locate must survive the pre order checks. A position side set in the
-        // properties wins over the holdings-based mapping.
-        private static IEnumerable<TestCaseData> LocateCleanupCases()
-        {
-            // Bloomberg carries the locate as raw fix tags
-            yield return new TestCaseData(CreateLocateBrokerTagProperties(), 0m, -10m, true)
-                .SetName("ShortSellKeepsLocateBroker");
-            yield return new TestCaseData(CreateLocateBrokerTagProperties(), 0m, 10m, false)
-                .SetName("BuyToOpenDropsLocateBroker");
-            yield return new TestCaseData(CreateLocateBrokerTagProperties(), 20m, -10m, false)
-                .SetName("SellToCloseDropsLocateBroker");
-            yield return new TestCaseData(CreateLocateBrokerTagProperties(), -10m, 10m, false)
-                .SetName("BuyToCloseDropsLocateBroker");
-            yield return new TestCaseData(CreateLocateBrokerTagProperties(), 100m, -300m, true)
-                .SetName("CrossZeroSellKeepsLocateBroker");
-            yield return new TestCaseData(CreateLocateTagProperties(), 0m, -10m, true)
-                .SetName("ShortSellKeepsLocateTags");
-            yield return new TestCaseData(CreateLocateTagProperties(), 20m, -10m, false)
-                .SetName("SellToCloseDropsLocateTags");
-
-            // TerminalLink keeps the locate broker and locate id together; either one alone counts
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateBroker = "MLCO", LocateId = "LOC-123" }, 0m, -10m, true)
-                .SetName("ShortSellKeepsTerminalLinkLocate");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateBroker = "MLCO", LocateId = "LOC-123" }, 100m, -300m, true)
-                .SetName("CrossZeroSellKeepsTerminalLinkLocate");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateBroker = "MLCO", LocateId = "LOC-123" }, 20m, -10m, false)
-                .SetName("SellToCloseDropsTerminalLinkLocate");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateId = "LOC-123" }, 0m, -10m, true)
-                .SetName("ShortSellKeepsLoneLocateId");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateId = "LOC-123" }, 20m, -10m, false)
-                .SetName("SellToCloseDropsLoneLocateId");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateBroker = "MLCO", LocateId = "LOC-123", PositionSide = OrderPosition.SellToOpen }, 20m, -10m, true)
-                .SetName("TerminalLinkPositionSideShortKeepsLocate");
-            yield return new TestCaseData(new TerminalLinkOrderProperties { LocateBroker = "MLCO", LocateId = "LOC-123", PositionSide = OrderPosition.SellToClose }, 0m, -10m, false)
-                .SetName("TerminalLinkPositionSideCloseDropsLocate");
-
-            // Wolverine carries the locate broker only
-            yield return new TestCaseData(new WolverineOrderProperties { LocateBroker = "MLCO", PositionSide = OrderPosition.SellToOpen }, 20m, -10m, true)
-                .SetName("WolverinePositionSideShortKeepsLocate");
-            yield return new TestCaseData(new WolverineOrderProperties { LocateBroker = "MLCO", PositionSide = OrderPosition.SellToClose }, 0m, -10m, false)
-                .SetName("WolverinePositionSideCloseDropsLocate");
-        }
-
-        [TestCaseSource(nameof(LocateCleanupCases))]
-        public void MarketOrderKeepsLocateOnlyOnShortSells(IOrderProperties properties, decimal holdings, decimal quantity, bool locateKept)
-        {
-            // Arrange
-            Security msft;
-            var algo = GetAlgorithm(out msft, 2m, 0);
-            Update(msft, 100m);
-            algo.Portfolio[Symbols.MSFT].SetHoldings(100m, holdings);
-            var locateBefore = GetLocateState(properties);
-
-            // Act
-            algo.MarketOrder(Symbols.MSFT, quantity, orderProperties: properties);
-
-            // Assert: the submitted request carries the locate only when the order opens a short
-            var submitted = ((SubmitOrderRequest)_fakeOrderProcessor.ProcessedOrdersRequests[1]).OrderProperties;
-            Assert.AreEqual(locateKept ? locateBefore : default, GetLocateState(submitted));
-
-            // The algorithm's own object is never touched
-            Assert.AreEqual(locateBefore, GetLocateState(properties));
-        }
-
-        // The locate fields of each supported properties type as one comparable value;
-        // all null means the order carries no locate.
-        private static (string LocateBroker, string LocateId, string LocateBrokerTag, string LocateReqdTag) GetLocateState(IOrderProperties properties)
-        {
-            return properties switch
-            {
-                TerminalLinkOrderProperties terminalLink => (terminalLink.LocateBroker, terminalLink.LocateId, null, null),
-                WolverineOrderProperties wolverine => (wolverine.LocateBroker, null, null, null),
-                FixOrderProperties fix => (null, null, fix.AdditionalProperties.GetValueOrDefault("5700"), fix.AdditionalProperties.GetValueOrDefault("114")),
-                _ => default
-            };
-        }
-
-        private static BloombergFixOrderProperties CreateLocateBrokerTagProperties()
-        {
-            var properties = new BloombergFixOrderProperties();
-            properties.AdditionalProperties["5700"] = "MLCO";
-            return properties;
-        }
-
-        private static BloombergFixOrderProperties CreateLocateTagProperties()
-        {
-            var properties = new BloombergFixOrderProperties();
-            properties.AdditionalProperties["114"] = "N";
-            properties.AdditionalProperties["5700"] = "MLCO";
-            return properties;
-        }
-
-        // Reusing one properties object across orders must not lose the locate for a later short:
-        // the cleanup may only touch the order it checked, never the algorithm's own object.
-        [Test]
-        public void MarketOrderLocateCleanupDoesNotMutateReusedProperties()
-        {
-            // Arrange
-            Security msft;
-            var algo = GetAlgorithm(out msft, 2m, 0);
-            Update(msft, 100m);
-            var properties = CreateLocateBrokerTagProperties();
-
-            // Act: close the long first, then open the short, both with the same properties object
-            algo.Portfolio[Symbols.MSFT].SetHoldings(100m, 10m);
-            algo.MarketOrder(Symbols.MSFT, -10, orderProperties: properties);
-            algo.Portfolio[Symbols.MSFT].SetHoldings(100m, 0m);
-            algo.MarketOrder(Symbols.MSFT, -10, orderProperties: properties);
-
-            // Assert: the algorithm's object still carries the locate for the short leg
-            Assert.AreEqual("MLCO", properties.AdditionalProperties["5700"]);
-
-            // The close leg went out without the locate, the short leg with it
-            var closeRequest = (SubmitOrderRequest)_fakeOrderProcessor.ProcessedOrdersRequests[1];
-            var shortRequest = (SubmitOrderRequest)_fakeOrderProcessor.ProcessedOrdersRequests[2];
-            Assert.IsFalse(((FixOrderProperties)closeRequest.OrderProperties).AdditionalProperties.ContainsKey("5700"));
-            Assert.AreEqual("MLCO", ((FixOrderProperties)shortRequest.OrderProperties).AdditionalProperties["5700"]);
-        }
-
         private static object[] LiquidateWorksAsExpectedTestCases =
         {
             new object[] { Language.CSharp, true, true, TimeInForce.Day },
